requirements.txt
# Satellite Image Super-Resolution - Requirements
# For the ML Track Hackathon

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
rasterio>=1.3.0

# Data & Utils
numpy>=1.24.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
streamlit>=1.28.0

# Metrics
lpips>=0.1.4

# Google Earth Engine (optional - for GEE data fetching)
# earthengine-api>=0.1.370

# Development
jupyter>=1.0.0
ipywidgets>=8.0.0

# For Colab (pre-installed, but listed for completeness)
# google-colab
